FX-3295 Separated SearchResults component
